fix CS0030: LayerMask needs an int hop before the uint cast
Session: f8b84966-e09a-4f21-8f40-80a79a641e7a

+++ net.yozolab.particletools/Editor/ParticleScrubberOverlay.cs
@@ -206,7 +206,7 @@ namespace YozoLab.ParticleTools
                         concatenated, InternalEditorUtility.layers);
                     if (EditorGUI.EndChangeCheck())
                     {
-                        BuiltinPreviewBridge.PreviewLayers = unchecked((uint)
+                        BuiltinPreviewBridge.PreviewLayers = unchecked((uint)(int)
                             InternalEditorUtility.ConcatenatedLayersMaskToLayerMask(newConcatenated));
                         ParticleScrubController.RequestResync();
                     }
